# Use MemoryExtensions.Split over string.Split to decrease allocations/improve perf