--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Annotations/Anchoring/TextSelectionProcessor.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Annotations/Anchoring/TextSelectionProcessor.cs
@@ -432,18 +432,18 @@ namespace MS.Internal.Annotations.Anchoring
         private static void GetLocatorPartSegmentValues(ContentLocatorPart locatorPart, int segmentNumber, out int startOffset, out int endOffset)
         {
             if (segmentNumber < 0)
-                throw new ArgumentException("segmentNumber");
+                throw new ArgumentException(nameof(segmentNumber));
 
-            string segmentString = locatorPart.NameValuePairs[SegmentAttribute + segmentNumber.ToString(NumberFormatInfo.InvariantInfo)];
+            ReadOnlySpan<char> segmentString = locatorPart.NameValuePairs[SegmentAttribute + segmentNumber.ToString(NumberFormatInfo.InvariantInfo)].AsSpan();
 
-            string[] values = segmentString.Split(Separator);
-            if (values.Length != 2)
+            Span<Range> splitRegions = stackalloc Range[3];
+            if (segmentString.Split(splitRegions, Separator) != 2)
             {
                 throw new ArgumentException(SR.Format(SR.InvalidLocatorPart, SegmentAttribute + segmentNumber.ToString(NumberFormatInfo.InvariantInfo)));
             }
 
-            startOffset = Int32.Parse(values[0], NumberFormatInfo.InvariantInfo);
-            endOffset = Int32.Parse(values[1], NumberFormatInfo.InvariantInfo);
+            startOffset = int.Parse(segmentString[splitRegions[0]], NumberFormatInfo.InvariantInfo);
+            endOffset = int.Parse(segmentString[splitRegions[1]], NumberFormatInfo.InvariantInfo);
         }
 
         /// <summary>