--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/ThemeDictionaryExtension.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/ThemeDictionaryExtension.cs
@@ -95,7 +95,7 @@ namespace System.Windows
 
         // Build the Uri for the assembly:
         // /AssemblyName;Component/themes/<CurrentTheme>.<Color>.xaml
-        private static Uri GenerateUri(string assemblyName, string resourceName, string themeName)
+        private static Uri GenerateUri(string assemblyName, string resourceName, ReadOnlySpan<char> themeName)
         {
             StringBuilder uri = new StringBuilder(assemblyName.Length + 50);
 
@@ -113,11 +113,12 @@ namespace System.Windows
             uri.Append(resourceName);
             uri.Append(".xaml");
 
-            return new System.Uri(uri.ToString(), System.UriKind.RelativeOrAbsolute);
+            return new Uri(uri.ToString(), System.UriKind.RelativeOrAbsolute);
         }
 
         internal static Uri GenerateFallbackUri(ResourceDictionary dictionary, string resourceName)
         {
+            Span<Range> splitRegions = stackalloc Range[3];
             for (int i = 0; i < _themeDictionaryInfos.Count; i++)
             {
                 ThemeDictionaryInfo info = _themeDictionaryInfos[i];
@@ -131,8 +132,11 @@ namespace System.Windows
                 }
                 if ((ResourceDictionary)info.DictionaryReference.Target == dictionary)
                 {
-                    string themeName = resourceName.Split('/')[1];
-                    return GenerateUri(info.AssemblyName, resourceName, themeName);
+                    ReadOnlySpan<char> nameSpan = resourceName.AsSpan();
+                    if (nameSpan.Split(splitRegions, '/') < 2)
+                        throw new ArgumentOutOfRangeException();
+
+                    return GenerateUri(info.AssemblyName, resourceName, nameSpan[splitRegions[1]]);
                 }
             }
             return null;