--- a/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/XamlSchemaContext.cs
+++ b/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/XamlSchemaContext.cs
@@ -171,19 +171,20 @@ namespace System.Xaml
             return result;
         }
 
-        string GetPrefixForClrNs(string clrNs, string assemblyName)
+        private string GetPrefixForClrNs(string clrNs, string assemblyName)
         {
             if (string.IsNullOrEmpty(assemblyName))
             {
                 return KnownStrings.LocalPrefix;
             }
 
-            var sb = new StringBuilder();
-            foreach (string segment in clrNs.Split('.'))
+            StringBuilder sb = new();
+            ReadOnlySpan<char> values = clrNs.AsSpan();
+            foreach (Range segment in values.Split('.'))
             {
-                if (!string.IsNullOrEmpty(segment))
+                if (!values[segment].IsEmpty)
                 {
-                    sb.Append(char.ToLower(segment[0], TypeConverterHelper.InvariantEnglishUS));
+                    sb.Append(char.ToLower(values[segment][0], TypeConverterHelper.InvariantEnglishUS));
                 }
             }