## Description

Picks off most places where `string.Split` is used just for parts count or where additional operations are performed on the split strings so it makes sense to work with `ReadOnlySpan<char>`/`SpanSplitEnumerator<char>` instead.

### Using `string.Split` vs `SpanSplitEnumerator<char>` for counting

| Method   | updatedFilter        | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |--------------------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | Image(...))\|\*.\* [68] |  49.09 ns |   0.676 ns |    0.600 ns | 0.0172 |         672 B |         288 B |
| PR__EDIT | Image(...))\|\*.\* [68] |  17.23 ns |   0.075 ns |    0.066 ns |      - |         412 B |             - |

<details><summary>Benchmark code</summary>

```csharp
[Benchmark]
[Arguments("Image Files(*.BMP;*.JPG;*.GIF)|*.BMP;*.JPG;*.GIF|All files (*.*)|*.*")]
public  int Original(string updatedFilter)
{
    string[] formats = updatedFilter.Split('|');

    return formats.Length; 
}

[Benchmark]
[Arguments("Image Files(*.BMP;*.JPG;*.GIF)|*.BMP;*.JPG;*.GIF|All files (*.*)|*.*")]
public int PR__EDIT(string updatedFilter)
{
    int formatsCount = 0;
    foreach (Range range in updatedFilter.AsSpan().Split('|'))
        formatsCount++;

    return formatsCount;
}
```

</details>

## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build/CI.

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9600)