@harshit7962 2025-01-24
Aah, I was toggling between old and new merge experience of merging PRs and did not realize that it changed the default back to commit merge 😅
Will change the default now.